general: add automated trade with zebras
Add a startTrades function, and a toggle for automatic trading. Generally,
zebras are the best race to trade with as they can give titanium.

Works by determining a max limit for gold/catpower to trade, and then
divides that by an extra percentage for each race. Currently set zebras to
trade at 25% of the requested value, so essentially "25% of the 50% max
resources it will spend"

This enables adding multiple races since the first could be set with a low
amount and the second with the remaining amount. No validation occurs but
the expectation is that trade amount for each race combined should always
be <= 1.0

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>