crafting: toggle limited field
Might be better to make this a text label that changes onclick, but I
couldn't see an easy way to change the label's text once it was created.
This is modeled after the trading per-seasons code. However, make it a
single checkbox, and increase the minWidth of all the options labels so
that these checkboxes always appear lined up for readability.

Also, fix the trading input.on function closure to use the passed option
instead of looking it up manually.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>