kitten-scientists: implement a simple activity system
This patch modifies the game so that we no longer print messages as we
perform actions. Instead, the user may click "show activity" in the
options list in order to print a display of all actions taken by the
scientists since they last clicked it. It conveniently also shows the
previous date as well.

Possibly this might be more valuable as a listing in a display box but I
couldn't figure out how to do that.

Instead, this just allows us to condense all our output until the user
wants to see it. In this way, after long absense from the game such as
over night or several days, the user can easily see what was acomplished.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>