getLowestCraftAmount: fix case where value available is 0
The way that we cacluated the lowest amount resulted in ignoring any times
when getValue returned zero. This results in ignoring stock limits when
your available amount is below the stock. This is obviously not a good
thing.

Fix it so that we use undefined instead of 0 as the inializer check. Also,
fix getLowestTradeAmount in a similar way, though it actually seems to
have had a slightly different method to avoid this same problem. We want
to be consistent either way though.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>