crafting: only store season for limiting if we actually crafted
Previously, we stored it every time that the craft triggered, which might
not be able to craft anything simply due to other restrictions. This meant
that limited goods could theoretically be blocked for a long period of
time. Fix this by only storing the season if we actually have any amount
to craft, not every time.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>