Trade: Remove per race trigger
We have a user configurable trade ratio but this is only checked for gold
and not the trade item itself.  In most cases this is fine as we'll be
above both ratios normally.  However for Leviathans we're unlikely to be
close enough to the trigger in Unobtanium and thus won't trade in that
short window.