Update contributors
Some of the prior contributors are no longer returned from the GitHub API. I'm not sure why that is, so I added them as static data.

Some of the contributors added in 644ab7aa82830dcabf69ece36cf22dd4bef6578c were duplicates of existing contributors or had missing links. This was corrected.