feat(village): New job distribution logic
The "limited" option has been removed as it didn't provide any meaningful change in behavior, compared to just setting a limit value.

The limit value can be set to -1, meaning unlimited, and the unlimited jobs will be filled when the limited jobs have reached their limit.
If you want kittens to be distributed to those jobs earlier, set a very high limit.

Kittens will always be assigned to the job with the least amount of kittens
currently assigned.