feat(core): Unified "trigger" concept
All options that have a trigger can now be managed through the same UI and API. Some use case specific prompts have been lost in the process. Those should be restored later in the documentation.

This also fixes remaining cases of the trigger icon not being rendered on some devices.