feat(core): Always enable save game manager
We no longer use our own internal dirty check and now just always install the save manager and save with the game.

The experimental engine state persistence is now always active. Loading is prevented for now.
The legacy storage is always written as well and is still used to restore previous settings.