feat(ui): Remove "Options" section
The last remaining option "Observe astro events" is now hosted in the "Science" section.

We might have an "Options" section again at a later time to control the meta behavior of KS. Everything relating to controlling the game should be hosted in a settings section relating to a tab in the game.