feat(core): Retain meta behavior on state change
This should ensure that the engine is correctly turned on or off, depending on the loaded state, and that the main checkbox also reflects that state.

Additionally, when loading a state through state management, your meta behavior should be retained. Meta behavior from the stored state will not override your local state. For most use cases, this should be much more intuitive.

Meta behavior is your KS internal interval and the state management itself.