feat(ui): Add a panel for KS internals
This will finally allow access to all internal settings of KS.
For the time being, the internal interval KS runs at has been exposed.
Be very careful when lowering this value! A too low value can freeze
your game. If this is persisted to your save, it will be very hard to
recover.