feat(time): Adjust default reset automation settings
Previously, pretty much any option in the Reset Timeline automation, was enabled by default, and set to infinite.
This indicates that you need an infinite amount of pretty much everything, for the game to be reset.

The new defaults are safe, in that they are disabled by default, but retain the infinite value. This makes it safe for the user to adjust either the amount or the enabled state first, without causing anything unexpected.

Fixes #183