fix(settings): avoid sharing references with engine
This fixes an issue that occurs when updating a state and then switching
to another one. The `this._host.engine.stateSerialize()` function will
return a reference that it itself is holding on to at some level, and
then `this.states` will contain that live reference. Further
modifications will be persisted, and in the worst case, multiple states
can end up containing the same reference if saved in the same session.

A simple repro:

1. Create new state A
2. Modify settings
3. Create new state B
4. Load state A

The changes made before saving B are applied to A. You can modify and
update A, but then on loading B the changes will be applied there as
well.

I opted to use `structuredClone` to fix this so that whenever we receive
or return a state, we make a clone of it. `structuredClone` has been
available since 2022, and in this case we could have also used a
round-trip through JSON stringify. I also created an abstraction that
maybe was a little overkill, but I wasn't confident that all consumers
of loading objects wouldn't hold on to references now or in the future.