feat: Drop snapshots/schemata
This was a fun idea to move KS settings editing into another domain. A lot was learned from this experiment. But it's ultimately too restrictive to development, sparked little interest, and is of little value.

Instead of hanging on to this failed feature, it will be discarded to allow focus on more useful features.