fix(core): Inconsistent trigger handling
Some of our triggers are expressed as percentage, some as an absolute value. Developers always had to carefully pick the correct handler for the given kind of trigger. This was cumbersome and error-prone.

We now introduce a new data structure to handle this kinds of triggers more intuitively and without explicit selection.

Fixes #627