feat(workshop): New crafting mechanics
This removes the logic that makes limited crafts unlimited under certain
conditions. The only exception is the manual ship override.

Limited crafts are now more complicated. See the comments in the
implementation for details. In summary, it tries to be less agressive in
crafting when you have very large differences between source materials
and target crafts. It also generalizes how base materials and crafted
resources are treated as source materials in limited crafting.

Closes #499