fix(core): Comments in UserScript code
Having the expanded source is useful to see a diff when installing an update of the userscript in some userscript managers.

The comments are really intended to help with the source code. Having the comments in the userscript might even give the impression that the userscript is the source code itself.

Fixes #843